implement TreedocDenseLocalList
