+++ client/src/crdt/list/deleting_mut_list.ts
@@ -3,6 +3,10 @@ import { Crdt } from "../core";
 import { Resettable } from "../helper_crdts";
 import { DeletingMutCSet } from "../set";
 import { MovableMutCList, MovableMutCListEntry } from "./movable_mut_list";
+import {
+  TreedocDenseLocalList,
+  TreedocLocWrapper,
+} from "./treedoc_dense_local_list";
 
 // TODO: reset.  It only actually works if DenseLocalList
 // has no tombstones.  Perhaps this is a reason to just
@@ -12,10 +16,10 @@ export class DeletingMutCList<C extends Crdt, InsertArgs extends any[]>
   extends MovableMutCList<
     C,
     InsertArgs,
-    TreedocLoc,
+    TreedocLocWrapper,
     DeletingMutCSet<
-      MovableMutCListEntry<C, TreedocLoc>,
-      [TreedocLoc, InsertArgs]
+      MovableMutCListEntry<C, TreedocLocWrapper>,
+      [TreedocLocWrapper, InsertArgs]
     >
   >
   implements Resettable
@@ -37,7 +41,9 @@ export class DeletingMutCList<C extends Crdt, InsertArgs extends any[]>
     // TODO: might throw error due to double-parent.
     // Should change both owns methods to guard against this
     // (return false on rootCrdt).
-    return this.set.owns(value.parent as MovableMutCListEntry<C, TreedocLoc>);
+    return this.set.owns(
+      value.parent as MovableMutCListEntry<C, TreedocLocWrapper>
+    );
   }
 
   reset() {
