+++ client/src/crdt/list/index.ts
@@ -1,4 +1,10 @@
+export * from "./abstract_list";
+export * from "./deleting_mut_list";
+export * from "./dense_local_list";
 export * from "./interfaces";
-// TODO
-export * from "./old_list";
-export * from "./movable_list";
+export * from "./list_from_map";
+export * from "./movable_mut_list";
+export * from "./primitive_list";
+export * from "./resetting_mut_list";
+export * from "./tombstone_mut_list";
+export * from "./treedoc_dense_local_list";
