--- client/src/crdt/list/old_list.proto
@@ -1,40 +1,0 @@
-message TreedocIdMessage {
-  required bytes path = 1;
-  repeated uint32 disIndices = 2 [packed=true];
-  repeated string disSendersIndex = 3;
-  repeated uint32 disSenders = 4 [packed=true];
-  repeated sint32 disUniqueNumbers = 5 [packed=true];
-}
-
-message PrimitiveListInsertMessage {
-  required bytes seqId = 1;
-  required bytes value = 2;
-}
-
-message PrimitiveListRangeMessage {
-  required bytes seqIdStart = 1;
-  required bytes values = 2;
-}
-
-message PrimitiveListDeleteMessage {
-  required bytes seqIdStart = 1;
-  optional bytes seqIdEnd = 2;
-}
-
-message PrimitiveListMessage {
-  oneof op {
-    PrimitiveListInsertMessage insert = 1;
-    PrimitiveListRangeMessage range = 2;
-    PrimitiveListDeleteMessage delete = 3;
-  }
-}
-
-message PrimitiveListEntrySave {
-  required bytes seqId = 1;
-  required uint32 senderCounter = 2;
-  required bytes value = 3;
-}
-
-message PrimitiveListSave {
-  repeated PrimitiveListEntrySave entries = 1;
-}
